requirements.txt
stable-baselines3[extra,tests,docs]>=0.10.0
box2d-py==2.3.8
pybullet
redis
gym-minigrid
scikit-optimize
optuna~=2.2.0
pyvirtualdisplay~=1.3.2
pytablewriter
seaborn
pyyaml>=5.1
sb3-contrib>=0.10.0
mysqlclient


gym~=0.17.3
matplotlib~=3.3.2
numpy~=1.18.1
torch~=1.7.0
ipython~=7.19.0
pillow~=8.0.1
sklearn~=0.0
scikit-learn~=0.23.2